+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added new Discord formatter supporting the Celestial Light Brigade's preferred Discord structure and layout for BGS reports.
 * Each faction now has its influence % shown in the on-screen activity window.
+* Factions are shown ordered by % influence, highest first in the on-screen activity window.
 
 ### Changes: